feat: チーム登録画面に利用規約・プライバシーポリシーの同意を追加 (issue #35)
- public/terms.txt, public/privacy.txt に規約文面をテキストファイルで保持
- チーム登録画面内のテキストエリアに規約文面を表示
- 利用規約・プライバシーポリシーへの同意チェックボックスを必須化
- app/register/page.tsx を Server Component に変更しテキストファイルを読み込み
- フォームロジックを app/register/register-form.tsx (Client Component) に分離
