+++ public/privacy.txt
@@ -1,0 +1,45 @@
+プライバシーポリシー
+
+やきゅスコ（以下「当サービス」）は、ユーザーの個人情報の取扱いについて、以下のとおりプライバシーポリシーを定めます。
+
+第1条（収集する情報）
+当サービスは以下の情報を収集します。
+(1) メールアドレス（アカウント認証のため）
+(2) チームID・チーム名（サービス提供のため）
+(3) 試合スコア・選手情報等、ユーザーが入力したデータ
+(4) アクセスログ（IPアドレス、ブラウザ情報等）
+
+第2条（利用目的）
+収集した情報は以下の目的で利用します。
+(1) アカウント認証およびサービス提供
+(2) サービスの品質改善
+(3) 重要なお知らせ・メンテナンス情報の通知
+(4) 法令に基づく対応
+
+第3条（第三者提供）
+当サービスは、以下の場合を除き、個人情報を第三者に提供しません。
+(1) ユーザー本人の同意がある場合
+(2) 法令に基づく場合
+(3) 人の生命・身体・財産の保護のため必要な場合
+
+第4条（業務委託）
+当サービスはサービス運営のため、以下のサービスを利用しています。これらのサービスプロバイダのプライバシーポリシーも合わせてご確認ください。
+- Supabase（データベース・認証）
+- Vercel（ホスティング）
+
+第5条（データの保管）
+ユーザーデータは安全なクラウド環境（Supabase）に保管されます。適切なセキュリティ対策を講じていますが、完全な安全性を保証するものではありません。
+
+第6条（データの削除）
+チームの削除を希望する場合は、サービス内のお問い合わせよりご連絡ください。リクエストを受けてから30日以内にデータを削除します。
+
+第7条（Cookieの利用）
+当サービスはセッション管理のためにCookieを使用します。ブラウザの設定によりCookieを無効にすることができますが、その場合サービスが正常に動作しない場合があります。
+
+第8条（お問い合わせ）
+個人情報の取扱いに関するお問い合わせは、サービス内のお問い合わせよりご連絡ください。
+
+第9条（改定）
+本ポリシーは必要に応じて改定されることがあります。重要な変更がある場合はサービス上でお知らせします。
+
+制定日: 2026年4月28日
