+++ public/terms.txt
@@ -1,0 +1,32 @@
+利用規約
+
+第1条（適用）
+本利用規約（以下「本規約」）は、やきゅスコ（以下「本サービス」）の利用に関する条件を定めるものです。ユーザーは本規約に同意の上、本サービスをご利用ください。
+
+第2条（利用登録）
+本サービスのチーム登録を完了した時点で、本規約に同意したものとみなします。登録者は18歳以上であること、または保護者の同意を得た上でご利用ください。
+
+第3条（サービスの内容）
+本サービスは、野球チームの試合スコア・打撃成績・投手成績を記録・管理するウェブアプリケーションです。本サービスは無料で提供しますが、予告なく内容を変更または終了する場合があります。
+
+第4条（禁止事項）
+ユーザーは以下の行為を行ってはなりません。
+(1) 法令または公序良俗に違反する行為
+(2) 本サービスの運営を妨害する行為
+(3) 他のユーザーまたは第三者の権利を侵害する行為
+(4) 不正アクセスまたはコンピュータウイルスの送信
+(5) その他、当サービスが不適切と判断する行為
+
+第5条（免責事項）
+本サービスは「現状のまま」提供されます。データの損失・消失について、当サービスは責任を負いません。定期的なバックアップはユーザー自身の責任で行ってください。
+
+第6条（サービスの変更・終了）
+当サービスは、事前の通知なく本サービスの内容を変更、または提供を終了することがあります。これによりユーザーに生じた損害について責任を負いません。
+
+第7条（規約の変更）
+当サービスは必要に応じて本規約を変更することがあります。変更後の規約はサービス上に掲示した時点で効力を生じます。
+
+第8条（準拠法・管轄裁判所）
+本規約は日本法に準拠します。本サービスに関する紛争は、東京地方裁判所を第一審の専属的合意管轄裁判所とします。
+
+制定日: 2026年4月28日
